app_examples/novel_v2/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.26.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
alembic>=1.13.0
